refactor(bottle): remove web-view directory
